refactor: flatten and reorganize client submodules
