+++ fireauth2/src/client/authorization/extra_params/mod.rs
@@ -1,0 +1,9 @@
+mod access_type;
+mod extra_param;
+mod include_granted_scopes;
+mod prompt;
+
+pub use access_type::*;
+pub use extra_param::*;
+pub use include_granted_scopes::*;
+pub use prompt::*;
