--- fireauth2/src/client/types/mod.rs
@@ -1,15 +1,0 @@
-mod authorization;
-mod exchange;
-mod extra_fields;
-mod extra_params;
-mod introspection;
-mod redirect;
-mod revocation;
-
-pub use authorization::*;
-pub use exchange::*;
-pub use extra_fields::*;
-pub use extra_params::*;
-pub use introspection::*;
-pub use redirect::*;
-pub use revocation::*;
